backend: factory refactor
 - move ApartmentFactory to own file outside of housing company.
   this is consistent splitting with other files
